Семинар 5 DZ3 через int

+++ seminars5/DZ3/Program.cs
@@ -1,18 +1,18 @@
-﻿double[] array = new double[6];
+﻿int[] array = new int[6];
   for (int i = 0; i < array.Length; i++ )
   {
     array[i] = new Random().Next(1, 100);
     Console.Write(array[i] + " ");
   }
 
-double maxNumber = array[0];
-double minNumber = array[0];
+int maxNumber = array[0];
+int minNumber = array[0];
 
   for (int i = 1; i < array.Length; i++)
   {
     if (maxNumber < array[i])
     {
-      maxNumber = array[i];
+        maxNumber = array[i];
     }
         if (minNumber > array[i])
     {
@@ -20,6 +20,6 @@ double minNumber = array[0];
     }
   }
 
-  double diff = maxNumber - minNumber;
+int diff = maxNumber - minNumber;
 
-  Console.WriteLine($"Разница между между максимальным ({maxNumber}) и минимальным({minNumber}) элементами: {diff}");
+Console.WriteLine($"Разница между между максимальным ({maxNumber}) и минимальным({minNumber}) элементами: {diff}");
